24309 РАВЕНСТВО

+++ rust/src/05/5347.rs
@@ -2,7 +2,7 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<i32>);
+    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<i64>);
     let mut input = || input.next().unwrap();
 
     for (a, b) in (0..input()).map(|_| (input(), input())) {
@@ -10,11 +10,11 @@ fn main() {
     }
 }
 
-fn get_lcm(a: i32, b: i32) -> i32 {
+fn get_lcm(a: i64, b: i64) -> i64 {
     a / get_gcd(a, b) * b
 }
 
-fn get_gcd(mut a: i32, mut b: i32) -> i32 {
+fn get_gcd(mut a: i64, mut b: i64) -> i64 {
     loop {
         if b == 0 {
             return a;
